Added debugging to Mass Storage demo. Also added in correct SENSE responses to sucessful and failed commands.
Correct data direction is now read from the CBW flags byte, and the correct enpoint selected. If a command fails with data still in one of the pipes, the pipe is now stalled.

Connecting the device gives a whole host of control transactions, followed by a bluescreen on XP.

Fixed Get/Clear Status/Feature Requests - now only the requests with the OTHER recipient of VENDOR type are handed to the user application.

git-svn-id: http://lufa-lib.googlecode.com/svn@92 d5102386-fcda-11dd-9fdb-3debd5008f28