Changed AudioOutput demo code to only perform the PWM timer config/deconfig when not using the LEDs display mode. Changed lower LED threshold for better audio visual when compiled in LED display mode.
Changed Mem_Free in DynAlloc to use the DEREF macro when setting the handle to NULL. Altered ALLOCABLE_BYTES macro to force type conversion to unsigned long to prevent overflows.

Lowered search time in USB bus reset workaround for powerless pullup devices from 100ms to 10ms - still ample for the detection of a 1ms USB frame.

Renamed USB_ALL_MODES macro to USB_CAN_BE_BOTH for clarity and consistancy.

Altered AVR_HOST_GetDeviceConfigDescriptor() so that the variable pointed to by the ConfigSizePtr parameter is not changed when a buffer is specified.

git-svn-id: http://lufa-lib.googlecode.com/svn@170 d5102386-fcda-11dd-9fdb-3debd5008f28