Added in inclusion of alloca header, with fallback alias for old avr-lib-c versions.
Moved Pipe_Freeze() out of the while loop in the host control transfer routine, as it only needed to be executed once and not once per loop iteration.

Added in DESCRIPTOR_TYPE, DESCRIPTOR_SIZE, DESCRIPTOR_CAST and DESCRIPTOR_PCAST convenience macros for host apps, and DESCRIPTOR_ADDRESS for device apps.

git-svn-id: http://lufa-lib.googlecode.com/svn@183 d5102386-fcda-11dd-9fdb-3debd5008f28