Added code to CDCHost demo to properly determine and set the notification interrupt endpoint frequency from the device descriptors. Fixed CDCHost header file macro guard name.
Cleaned up CDCHost and MassStorageHost demos by casting the endpoint descriptors to a pointer to a endpoint descriptor struture once and reading out values from the pointer rather than casting each time a value was needed.

Moved MagStripe driver out to the MagStripe demo folder, rather than the inside the library itself.

Re-added StillImageHost demo to the library. Updated code to function using the latest library API.

Added C linkage directives for when a C++ compiler is used with the library, adding C++ compatibility to the project.

Fixed off-by-one error when checking if the Usage Stack is full in the HID report parser.

git-svn-id: http://lufa-lib.googlecode.com/svn@273 d5102386-fcda-11dd-9fdb-3debd5008f28